+++ src/components/button/button.test.js
@@ -8,9 +8,10 @@ import {
 
 import { icon, tooltipPreview } from "../../../cypress/locators";
 import { positionOfElement, keyCode } from "../../../cypress/support/helper";
+import { CHARACTERS } from "../../../cypress/support/component-helper/constants";
 import CypressMountWithProviders from "../../../cypress/support/component-helper/cypress-mount";
 
-const testData = ["mp150ú¿¡üßä", "!@#$%^*()_+-=~[];:.,?{}&\"'<>"];
+const testData = [CHARACTERS.DIACRITICS, CHARACTERS.SPECIALCHARACTERS];
 
 const ButtonDifferentTypes = ({ ...props }) => {
   return (